lib/private.h: rename to atom.h
Makes sense to have it named like this now.
Would have been a bit nicer to do it earlier, but now should be fine too.
