+++ src/lib/Makefile.am
@@ -9,7 +9,7 @@ noinst_LTLIBRARIES = libfixedpoint.la
 libfixedpoint_la_SOURCES = fixedpoint.h fixedpoint.c
 
 liblldpctl_la_SOURCES = \
-	lldpctl.h private.h errors.c connection.c atom.c helpers.c helpers.h \
+	lldpctl.h atom.h errors.c connection.c atom.c helpers.c helpers.h \
 	atom-private.c \
 	atoms/config.c atoms/dot1.c atoms/dot3.c \
 	atoms/interface.c atoms/med.c atoms/mgmt.c atoms/port.c
