+++ src/lib/atom.h
